+++ Cargo.toml
@@ -99,6 +99,7 @@ members = [
     "crates/rpc/rpc-testing-util/",
     "crates/rpc/rpc-types-compat/",
     "crates/rpc/rpc/",
+    "crates/scroll/primitives",
     "crates/scroll/revm",
     "crates/stages/api/",
     "crates/stages/stages/",
@@ -401,6 +402,7 @@ reth-rpc-eth-types = { path = "crates/rpc/rpc-eth-types", default-features = fal
 reth-rpc-layer = { path = "crates/rpc/rpc-layer" }
 reth-rpc-server-types = { path = "crates/rpc/rpc-server-types" }
 reth-rpc-types-compat = { path = "crates/rpc/rpc-types-compat" }
+reth-scroll-primitives = { path = "crates/scroll/primitives" }
 reth-scroll-revm = { path = "crates/scroll/revm" }
 reth-stages = { path = "crates/stages/stages" }
 reth-stages-api = { path = "crates/stages/api" }
